Add Megaraid_sas support
Adding this driver support will allow users that use megaraid sas card the ability to install and use OE. RAID functionalities supported by these cards are at the hardware level and no other packages are required.

User mladen.perisic84 has tested both a modified 6.95.1 (7.0 beta 1) version of OE Enabling this on the current 7.0 branch will allow for easier upgrading for users that have built modified versions of the build. This user and other users that have this type of hardware and wish to use it will have the ability to.