RFC: Add ODROID-U2/U3 support
It's the minimal changes needed in LE to add support for ODROID-U2/U3 properly.

It's a major cleanup of all code floating around with a bunch of bugs fixed.
I will of course maintain this portage and offer support if it's merged.

It's a preliminary work to support more recent hardkernel hardware cleanly
and more easily.

Once the RFC is reviewed, the patch will be broken into small changeset for merging.

Signed-off-by: Jérôme Benoit <jerome.benoit@piment-noir.org>