Update tv_grab_file
Well, after a bunch of attempts, this is the simplest way i found to manage all formats recognized by busybox version of the zcat command. Not exactly my original goal (keep script simplest), but at least works without breaking the compatibility with the original behavior. I tested it with .gz, .bz2 and uncompressed files. Not tested with .xv files.