rpi: rpi4: add canupdate.sh to only install on compatible hardware
Prior to installing an image, confirm that the image is meant for an RPi,
and that the cpu is a BCM2838.

/proc/cpuinfo contains a "Revision:" field which is populated with
different information about the RPi device (manufacturer, cpu, etc.).

This field is only available through cpuinfo on 32-bit arm. 64-bit arm
devices need to access the field through the device tree, where it is stored
as hex. This information is available at the same location on 32-bit arm too.

Example field from cpuinfo: c03111
Example field from device-tree (with hexdump formatting): 00c03111

The significant digit for the cpu identification is the third (c0_3_111). For
a BCM2838, this digit will equal 3.

Signed-off-by: Ian Leonard <antonlacon@gmail.com>