busybox: init: fix toram when SYSTEM_IMAGE is with path
For the SYSTEM copy to /dev i just replaced / with _ to fix it.
By choice i did not just use the filename, so one can see which
system image is loaded if the path is e.g. /boot/le/12/Generic/SYSTEM.
As far as i can tell all the rest of init is fine with BOOT_IMAGE and
SYSTEM_IMAGE having slashes in there.
Just toram was broken.