connman: ignore kernel configured netdev and improve resolv.conf handling
When netbooting the device used for /flash and/or /storage NFS mounts
is already configured by the kernel and has to be ignored by connman.

Connman only autodetects that case for standard netboot with nfsroot=
on kernel command line but doesn't know about boot= or disk= netboot
in LibreELEC so we have to set up the ignore manually.

When using kernel network (auto-)configuration the kernel provides
/proc/net/pnp with nameserver and in case of dhcp autoconfiguration
also with domain info, in standard resolv.conf format.

Though connman parses /proc/net/pnp if present it only uses the
nameserver info in it when creating resolv.conf but not the domain
info. To fix that use /proc/net/pnp directly as resolv.conf in case
of netboot.

To handle corner cases not covered by connman or kernel resolv.conf
setup allow users to provide their own resolv.conf file in
/storage/.config/

Signed-off-by: Matthias Reichl <hias@horus.com>