libretro-mame2016: fix cross compilation
Replace the current patch with the same approach as libretro-mame:

Don't set PTR64=0 on 32bit archs, instead set ARCHITECTURE to an empty
string. That way we don't need to patch out hardcoded -m32 arguments.

While at it, disable the bgfx hw renderers, or that pile of build crap
tries to include X11 headers.