restore: adjust mtime when restore finishes if clock not set at start
A system restore could start before the system clock has been set by
NTP in the case of a faulty or missing RTC. This change checks if
the system year matches the epoch year, and, if true, adjusts mtime
at the end of the restore process. This provides the duration of
the restore to complete for the system clock to be set.

Signed-off-by: Ian Leonard <antonlacon@gmail.com>