image: calculate number of tracks based on SYSTEM_SIZE
In case `SYSTEM_SIZE` is >= 1024, `mcopy`/`mmd` fail to copy files /
create folders in the FAT32 partition with a message

     The devil is in the details: zero number of heads or sectors

The image is still created, but the files/folders are missing in the
image. Another workaround is adding

     export MTOOLS_SKIP_CHECK=1

before any mtools commands are called. In such case correct image is
created.

I added a check to verify, that `SYSTEM_SIZE` is multiple of 32, so there
are no odd results.